pyproject.toml
[project]
name = "apisocial"
version = "0.1.0"
description = ""
authors = [
    {name = "Kredo15",email = "kredo1511@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "setuptools (>=80.9.0,<81.0.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
    "faker (>=37.4.0,<38.0.0)",
    "authlib (>=1.6.1,<2.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "celery (>=5.5.3,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
